requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
multidict==6.0.5
openai==0.28.0
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pytesseract==0.3.10
python-docx==1.1.2
python-dotenv==1.0.1
reportlab==4.2.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
